moved LP components into sub dir
