+++ app/react/components/LandingPage/Team.js
